refactor: CollapsibleCard into it's own component
